# ui/src/openapi/openapi.json
  {"openapi":"3.1.0","info":{"title":"OpenApi specification for CurricuLink","description":"API documentation for the File Manager App","termsOfService":"Terms of service","contact":{"name":"Cesar Dillon","url":"https://github.com/EDillon05","email":"cesar.dillon1275@alumnos.udg.mx"},"license":{"name":"MIT"},"version":"1.1"},"servers":[{"url":"http://localhost:8080","description":"Local server"},{"url":"https://api.dilon.com","description":"Production server"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"General Info","description":"General Info API"}],"paths":{"/teachings":{"put":{"tags":["teaching-controller"],"operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeachingRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}},"post":{"tags":["teaching-controller"],"operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeachingRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/experiences":{"put":{"tags":["experience-controller"],"operationId":"update_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExperienceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}},"post":{"tags":["experience-controller"],"operationId":"save_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExperienceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/careers":{"put":{"tags":["General Info"],"operationId":"update_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CareerRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}},"post":{"tags":["General Info"],"operationId":"saveCareer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CareerRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/teachings/{teaching-id}":{"get":{"tags":["teaching-controller"],"operationId":"findById","parameters":[{"name":"teaching-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeachingResponse"}}}}}},"delete":{"tags":["teaching-controller"],"operationId":"deleteById","parameters":[{"name":"teaching-id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/teachings/search":{"get":{"tags":["teaching-controller"],"operationId":"searchByFilters","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}},{"name":"type","in":"query","required":false,"schema":{"type":"string"}},{"name":"startedAt","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"finishedAt","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseTeachingResponse"}}}}}}},"/teachings/owner":{"get":{"tags":["teaching-controller"],"operationId":"findAllByOwner","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseTeachingResponse"}}}}}}},"/experiences/{experience-id}":{"get":{"tags":["experience-controller"],"operationId":"findById_1","parameters":[{"name":"experience-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExperienceResponse"}}}}}},"delete":{"tags":["experience-controller"],"operationId":"deleteById_1","parameters":[{"name":"experience-id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/experiences/search":{"get":{"tags":["experience-controller"],"operationId":"searchByFilters_1","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}},{"name":"type","in":"query","required":false,"schema":{"type":"string"}},{"name":"startedAt","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"finishedAt","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseExperienceResponse"}}}}}}},"/experiences/owner":{"get":{"tags":["experience-controller"],"operationId":"findAllByOwner_1","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseExperienceResponse"}}}}}}},"/careers/{career-id}":{"get":{"tags":["General Info"],"operationId":"findById_2","parameters":[{"name":"career-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CareerResponse"}}}}}},"delete":{"tags":["General Info"],"operationId":"deleteById_2","parameters":[{"name":"career-id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/careers/search":{"get":{"tags":["General Info"],"operationId":"searchByFilters_2","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}},{"name":"type","in":"query","required":false,"schema":{"type":"string"}},{"name":"startedAt","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"finishedAt","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseCareerResponse"}}}}}}},"/careers/owner":{"get":{"tags":["General Info"],"operationId":"findAllByOwner_2","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseCareerResponse"}}}}}}},"/auth":{"get":{"tags":["Authentication"],"operationId":"findAll","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}}},"/auth/{user-id}":{"get":{"tags":["Authentication"],"operationId":"findById_3","parameters":[{"name":"user-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/auth/activate-account":{"get":{"tags":["Authentication"],"operationId":"confirm","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"TeachingRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","minLength":1},"type":{"type":"string","minLength":1},"description":{"type":"string","minLength":1},"country":{"type":"string","minLength":1},"startedAt":{"type":"string","format":"date-time"},"finishedAt":{"type":"string","format":"date-time"},"comment":{"type":"string"},"url":{"type":"string"},"status":{"type":"string","minLength":1},"assessmentName":{"type":"string","minLength":1},"programName":{"type":"string","minLength":1},"level":{"type":"string","minLength":1},"institution":{"type":"string","minLength":1},"modality":{"type":"string","minLength":1},"weeklyHours":{"type":"string","minLength":1}},"required":["assessmentName","country","description","finishedAt","institution","level","modality","name","programName","startedAt","status","type","weeklyHours"]},"ExperienceRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","minLength":1},"type":{"type":"string","minLength":1},"description":{"type":"string","minLength":1},"country":{"type":"string","minLength":1},"startedAt":{"type":"string","format":"date-time"},"finishedAt":{"type":"string","format":"date-time"},"comment":{"type":"string"},"url":{"type":"string"},"status":{"type":"string","minLength":1},"companyName":{"type":"string","minLength":1},"position":{"type":"string","minLength":1},"area":{"type":"string","minLength":1},"contractType":{"type":"string","minLength":1}},"required":["area","companyName","contractType","country","description","finishedAt","name","position","startedAt","status","type"]},"CareerRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","minLength":1},"type":{"type":"string","minLength":1},"description":{"type":"string","minLength":1},"country":{"type":"string","minLength":1},"startedAt":{"type":"string","format":"date-time"},"finishedAt":{"type":"string","format":"date-time"},"comment":{"type":"string"},"url":{"type":"string"},"status":{"type":"string","minLength":1},"level":{"type":"string","minLength":1},"institution":{"type":"string","minLength":1},"certification_name":{"type":"string","minLength":1},"licenseId":{"type":"string","minLength":1},"modality":{"type":"string","minLength":1},"graduation_date":{"type":"string","format":"date-time"}},"required":["certification_name","country","description","finishedAt","graduation_date","institution","level","licenseId","modality","name","startedAt","status","type"]},"RegisterRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":2147483647,"minLength":1},"lastName1":{"type":"string","maxLength":2147483647,"minLength":1},"lastName2":{"type":"string","maxLength":2147483647,"minLength":1},"email":{"type":"string"},"password":{"type":"string","maxLength":2147483647,"minLength":8}},"required":["email","lastName1","name","password"]},"AuthenticationRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":8}},"required":["password"]},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}},"TeachingResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"type":{"type":"string"},"description":{"type":"string"},"country":{"type":"string"},"startedAt":{"type":"string","format":"date-time"},"finishedAt":{"type":"string","format":"date-time"},"comment":{"type":"string"},"url":{"type":"string"},"status":{"type":"string"},"assessmentName":{"type":"string"},"programName":{"type":"string"},"level":{"type":"string"},"institution":{"type":"string"},"modality":{"type":"string"},"weeklyHours":{"type":"string"}}},"PageResponseTeachingResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/TeachingResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"ExperienceResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"type":{"type":"string"},"description":{"type":"string"},"country":{"type":"string"},"startedAt":{"type":"string","format":"date-time"},"finishedAt":{"type":"string","format":"date-time"},"comment":{"type":"string"},"url":{"type":"string"},"status":{"type":"string"},"companyName":{"type":"string"},"position":{"type":"string"},"area":{"type":"string"},"contractType":{"type":"string"}}},"PageResponseExperienceResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/ExperienceResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"CareerResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"type":{"type":"string"},"description":{"type":"string"},"country":{"type":"string"},"startedAt":{"type":"string","format":"date-time"},"finishedAt":{"type":"string","format":"date-time"},"comment":{"type":"string"},"url":{"type":"string"},"status":{"type":"string"},"level":{"type":"string"},"institution":{"type":"string"},"certification_name":{"type":"string"},"licenseId":{"type":"string"},"modality":{"type":"string"},"graduation_date":{"type":"string","format":"date-time"}}},"PageResponseCareerResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/CareerResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"lastName1":{"type":"string"},"lastName2":{"type":"string"},"email":{"type":"string"},"createdAt":{"type":"string"},"lastModifiedAt":{"type":"string"},"roles":{"type":"string"},"accountLocked":{"type":"boolean"},"enabled":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}